+++ CHANGELOG.md
@@ -1,4 +1,14 @@
 
+## [v0.3.0] - 2023-10-02
+
+### Changed
+
+- Added gRPC classes for Push C2
+- Added C2 RPC calls for hosting files
+- Added PayloadType RPC calls for parsing TypedArray values
+- Added TypedArray values for Build, Command, and C2 parameters
+- Updated ProxyStart/ProxyStop commands to take an optional local_port of 0 and have it dynamically chosen
+
 ## [v0.2.11-rc03] - 2023-06-08
 
 ### Changed